Move CodeActionCreateAnalyzer to the new location.
