--- src/Roslyn.Diagnostics.Analyzers/CSharp/CSharpDoNotUseGenericCodeActionCreateToCreateCodeAction.Fixer.cs
@@ -1,20 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-                                 
-using System.Composition;
-using System.Diagnostics;  
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace Roslyn.Diagnostics.Analyzers
-{                                 
-    /// <summary>
-    /// RS0005: Do not use generic CodeAction.Create to create CodeAction
-    /// </summary>
-    [ExportCodeFixProvider(LanguageNames.CSharp), Shared]
-    public class CSharpDoNotUseGenericCodeActionCreateToCreateCodeActionFixer : DoNotUseGenericCodeActionCreateToCreateCodeActionFixer
-    { 
-
-    }
-}
