+++ src/Roslyn.Diagnostics.Analyzers/Core/DoNotUseGenericCodeActionCreateToCreateCodeAction.cs
@@ -1,38 +1,106 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Immutable;
-using Microsoft.CodeAnalysis.Diagnostics;
+using System.Linq;
+using System.Threading;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Roslyn.Diagnostics.Analyzers
-{                   
-    /// <summary>
-    /// RS0005: Do not use generic CodeAction.Create to create CodeAction
-    /// </summary>
-    public abstract class DoNotUseGenericCodeActionCreateToCreateCodeActionAnalyzer : DiagnosticAnalyzer
+{
+    public abstract class CodeActionCreateAnalyzer<TLanguageKindEnum> : DiagnosticAnalyzer where TLanguageKindEnum : struct
     {
-        internal const string RuleId = "RS0005";
+        internal const string CodeActionMetadataName = "Microsoft.CodeAnalysis.CodeActions.CodeAction";
+        internal const string CreateMethodName = "Create";
 
         private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotUseGenericCodeActionCreateToCreateCodeActionTitle), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
-        
         private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotUseGenericCodeActionCreateToCreateCodeActionMessage), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
         private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(RoslynDiagnosticsAnalyzersResources.DoNotUseGenericCodeActionCreateToCreateCodeActionDescription), RoslynDiagnosticsAnalyzersResources.ResourceManager, typeof(RoslynDiagnosticsAnalyzersResources));
-        
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                             s_localizableTitle,
-                                                                             s_localizableMessage,
-                                                                             DiagnosticCategory.Performance,
-                                                                             DiagnosticSeverity.Warning,
-                                                                             isEnabledByDefault: false,
-                                                                             description: s_localizableDescription,
-                                                                             helpLinkUri: null,     // TODO: add MSDN url
-                                                                             customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
+        internal static readonly DiagnosticDescriptor DontUseCodeActionCreateRule = new DiagnosticDescriptor(
+            RoslynDiagnosticIds.DontUseCodeActionCreateRuleId,
+            s_localizableTitle,
+            s_localizableMessage,
+            "Performance",
+            DiagnosticSeverity.Warning,
+            isEnabledByDefault: true,
+            description: s_localizableDescription,
+            customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override void Initialize(AnalysisContext analysisContext)
-        { 
-            
+        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics
+        {
+            get { return ImmutableArray.Create(DontUseCodeActionCreateRule); }
+        }
+
+        public override void Initialize(AnalysisContext context)
+        {
+            context.RegisterCompilationStartAction(CreateAnalyzerWithinCompilation);
+        }
+
+        private void CreateAnalyzerWithinCompilation(CompilationStartAnalysisContext context)
+        {
+            context.CancellationToken.ThrowIfCancellationRequested();
+
+            var codeActionSymbol = context.Compilation.GetTypeByMetadataName(CodeActionMetadataName);
+            if (codeActionSymbol == null)
+            {
+                return;
+            }
+
+            var createSymbols = codeActionSymbol.GetMembers(CreateMethodName).Where(m => m is IMethodSymbol);
+            if (createSymbols == null)
+            {
+                return;
+            }
+
+            var createSymbolsSet = ImmutableHashSet.CreateRange(createSymbols);
+            context.RegisterCodeBlockStartAction<TLanguageKindEnum>(GetCodeBlockStartedAnalyzer(createSymbolsSet).CreateAnalyzerWithinCodeBlock);
+        }
+
+        protected abstract AbstractCodeBlockStartedAnalyzer GetCodeBlockStartedAnalyzer(ImmutableHashSet<ISymbol> symbols);
+
+        protected abstract class AbstractCodeBlockStartedAnalyzer
+        {
+            private readonly ImmutableHashSet<ISymbol> _symbols;
+
+            public AbstractCodeBlockStartedAnalyzer(ImmutableHashSet<ISymbol> symbols)
+            {
+                _symbols = symbols;
+            }
+
+            protected abstract void GetSyntaxAnalyzer(CodeBlockStartAnalysisContext<TLanguageKindEnum> context, ImmutableHashSet<ISymbol> symbols);
+
+            public void CreateAnalyzerWithinCodeBlock(CodeBlockStartAnalysisContext<TLanguageKindEnum> context)
+            {
+                GetSyntaxAnalyzer(context, _symbols);
+            }
+        }
+
+        protected abstract class AbstractSyntaxAnalyzer
+        {
+            private readonly ImmutableHashSet<ISymbol> _symbols;
+
+            public AbstractSyntaxAnalyzer(ImmutableHashSet<ISymbol> symbols)
+            {
+                _symbols = symbols;
+            }
+
+            private bool IsCodeActionCreate(SyntaxNode expression, SemanticModel semanticModel, CancellationToken cancellationToken)
+            {
+                var symbolInfo = semanticModel.GetSymbolInfo(expression, cancellationToken);
+                return symbolInfo.Symbol != null && _symbols.Contains(symbolInfo.Symbol);
+            }
+
+            protected void AnalyzeInvocationExpression(SyntaxNode name, SemanticModel semanticModel, Action<Diagnostic> addDiagnostic, CancellationToken cancellationToken)
+            {
+                if (!IsCodeActionCreate(name, semanticModel, cancellationToken))
+                {
+                    return;
+                }
+
+                addDiagnostic(Diagnostic.Create(DontUseCodeActionCreateRule, name.Parent.GetLocation()));
+            }
         }
     }
 }
