--- src/Roslyn.Diagnostics.Analyzers/UnitTests/DoNotUseGenericCodeActionCreateToCreateCodeActionTests.Fixer.cs
@@ -1,33 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.CodeAnalysis.UnitTests;
-using Xunit;
-
-namespace Roslyn.Diagnostics.Analyzers.UnitTests
-{
-    public class DoNotUseGenericCodeActionCreateToCreateCodeActionFixerTests : CodeFixTestBase
-    {
-        protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
-        {
-            return new BasicDoNotUseGenericCodeActionCreateToCreateCodeActionAnalyzer();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new CSharpDoNotUseGenericCodeActionCreateToCreateCodeActionAnalyzer();
-        }
-
-        protected override CodeFixProvider GetBasicCodeFixProvider()
-        {
-            return new BasicDoNotUseGenericCodeActionCreateToCreateCodeActionFixer();
-        }
-
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
-        {
-            return new CSharpDoNotUseGenericCodeActionCreateToCreateCodeActionFixer();
-        }
-    }
-}
