--- src/Roslyn.Diagnostics.Analyzers/VisualBasic/BasicDoNotUseGenericCodeActionCreateToCreateCodeAction.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Roslyn.Diagnostics.Analyzers     
-    ''' <summary>
-    ''' RS0005: Do not use generic CodeAction.Create to create CodeAction
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicDoNotUseGenericCodeActionCreateToCreateCodeActionFixer
-        Inherits DoNotUseGenericCodeActionCreateToCreateCodeActionFixer 
-
-    End Class
-End Namespace
