+++ src/Roslyn.Diagnostics.Analyzers/VisualBasic/Roslyn.Diagnostics.VisualBasic.Analyzers.vbproj
@@ -102,7 +102,6 @@
     <Compile Include="BasicInvokeTheCorrectPropertyToEnsureCorrectUseSiteDiagnostics.vb" />
     <Compile Include="BasicInvokeTheCorrectPropertyToEnsureCorrectUseSiteDiagnostics.Fixer.vb" />
     <Compile Include="BasicDoNotUseGenericCodeActionCreateToCreateCodeAction.vb" />
-    <Compile Include="BasicDoNotUseGenericCodeActionCreateToCreateCodeAction.Fixer.vb" />
     <Compile Include="BasicOverrideObjectEqualsObjectWhenImplementingIequatable.vb" />
     <Compile Include="BasicOverrideObjectEqualsObjectWhenImplementingIequatable.Fixer.vb" />
     <Compile Include="BasicDoNotInvokeDiagnosticDescriptor.vb" />
